# backend/api-spec.yaml
openapi: 3.0.0
info:
  title: AI4Devs Candidate API
  version: 1.0.0
  description: API for managing candidate data in the AI4Devs recruitment system.
paths:
  /candidates:
    post:
      summary: Add a new candidate
      description: Adds a new candidate to the system and returns the created candidate data.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                  description: First name of the candidate
                  minLength: 2
                  maxLength: 50
                  pattern: '^[a-zA-ZñÑáéíóúÁÉÍÓÚ ]+$'
                lastName:
                  type: string
                  description: Last name of the candidate
                  minLength: 2
                  maxLength: 50
                  pattern: '^[a-zA-ZñÑáéíóúÁÉÍÓÚ ]+$'
                email:
                  type: string
                  description: Email address of the candidate
                  pattern: '^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$'
                phone:
                  type: string
                  description: Phone number of the candidate
                  pattern: '^\+?\d{1,3}?[- .]?\(?(?:\d{2,3})\)?[- .]?\d\d\d[- .]?\d\d\d\d$'
                address:
                  type: string
                  description: Address of the candidate
                  maxLength: 100
                educations:
                  type: array
                  description: Education history of the candidate
                  items:
                    type: object
                    properties:
                      institution:
                        type: string
                        description: Institution name
                        maxLength: 100
                      title:
                        type: string
                        description: Degree or title obtained
                        maxLength: 100
                      startDate:
                        type: string
                        description: Start date of the education
                        pattern: '^\d{4}-\d{2}-\d{2}$'
                      endDate:
                        type: string
                        description: End date of the education
                        pattern: '^\d{4}-\d{2}-\d{2}$'
                workExperiences:
                  type: array
                  description: Work experience of the candidate
                  items:
                    type: object
                    properties:
                      company:
                        type: string
                        description: Company name
                        maxLength: 100
                      position:
                        type: string
                        description: Position held
                        maxLength: 100
                      description:
                        type: string
                        description: Description of the job responsibilities
                        maxLength: 200
                      startDate:
                        type: string
                        description: Start date of the job
                        pattern: '^\d{4}-\d{2}-\d{2}$'
                      endDate:
                        type: string
                        description: End date of the job
                        pattern: '^\d{4}-\d{2}-\d{2}$'
                cv:
                  type: object
                  description: CV file information
                  properties:
                    filePath:
                      type: string
                      description: Path to the CV file
                    fileType:
                      type: string
                      description: Type of the CV file
      responses:
        '201':
          description: Candidate created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    description: Unique identifier for the candidate
                  firstName:
                    type: string
                  lastName:
                    type: string
                  email:
                    type: string
                  phone:
                    type: string
                  address:
                    type: string
                  educations:
                    type: array
                    items:
                      type: object
                      properties:
                        institution:
                          type: string
                        title:
                          type: string
                        startDate:
                          type: string
                        endDate:
                          type: string
                  workExperiences:
                    type: array
                    items:
                      type: object
                      properties:
                        company:
                          type: string
                        position:
                          type: string
                        description:
                          type: string
                        startDate:
                          type: string
                        endDate:
                          type: string
                  cv:
                    type: object
        '400':
          description: Bad request (invalid input data)
        '500':
          description: Internal server error
  
  /candidates/{id}/stage:
    put:
      summary: Update candidate stage
      description: Updates the stage of a candidate in the selection process
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the candidate to update
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - stage
              properties:
                stage:
                  type: string
                  description: The new stage to set for the candidate
                notes:
                  type: string
                  description: Optional notes about the stage change
      responses:
        '200':
          description: Stage updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  candidate_id:
                    type: integer
                  previous_stage:
                    type: string
                  new_stage:
                    type: string
                  updated_by:
                    type: string
                  updated_at:
                    type: string
                    format: date-time
        '400':
          description: Bad request (invalid input data)
        '404':
          description: Candidate not found
        '500':
          description: Internal server error
          
  /positions/{id}/candidates:
    get:
      summary: Get candidates for a position
      description: Retrieves a list of candidates for a specific position with sorting, filtering, and pagination
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the position
          schema:
            type: integer
        - name: sort
          in: query
          required: false
          description: Field to sort by
          schema:
            type: string
            enum: [name, score, stage]
            default: name
        - name: order
          in: query
          required: false
          description: Sort order (ascending or descending)
          schema:
            type: string
            enum: [asc, desc]
            default: asc
        - name: page
          in: query
          required: false
          description: Page number for pagination
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          required: false
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: stage
          in: query
          required: false
          description: Filter by interview stage
          schema:
            type: string
        - name: minScore
          in: query
          required: false
          description: Minimum score filter
          schema:
            type: number
            minimum: 0
            maximum: 10
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        candidate_id:
                          type: integer
                        candidate_name:
                          type: string
                        current_interview_step:
                          type: string
                        average_score:
                          type: number
                          nullable: true
                        last_interview_date:
                          type: string
                          format: date-time
                          nullable: true
                  pagination:
                    type: object
                    properties:
                      total:
                        type: integer
                      page:
                        type: integer
                      limit:
                        type: integer
                      pages:
                        type: integer
        '400':
          description: Bad request (invalid input parameters)
        '404':
          description: Position not found
        '500':
          description: Internal server error
  
  /upload:
    post:
      summary: Upload a file
      description: Uploads a file to the server. Only PDF and DOCX files are allowed.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: File to be uploaded
      responses:
        '200':
          description: File uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  filePath:
                    type: string
                    description: Path where the file is stored
                  fileType:
                    type: string
                    description: MIME type of the uploaded file
        '400':
          description: Invalid file type, only PDF and DOCX are allowed
        '500':
          description: Error during the file upload process